feat: type checking
